+++ RELEASES.md
@@ -22,6 +22,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * service-templates-core: TemplateQuery: fixed issue in recursive traversal query
 * service-templates-core: fixed issues, while generating screenshots for book
 * prov-template-compiler: PAST expression, detecting overloaded methods
+* prov-template-compiler: Python emitter, making use of overload annotation to disambiguate overloaded methods
 
 
 `2.2.2`